Change the microservice name ip_manager to private_ip_manager
