--- services/ip_manager/src/main/java/com/futurewei/alcor/ipmanager/http/status/Ipv4AddrStateInvalidException.java
@@ -1,8 +1,0 @@
-package com.futurewei.alcor.ipmanager.http.status;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(code= HttpStatus.BAD_REQUEST, reason="Ipv4 address state invalid")
-public class Ipv4AddrStateInvalidException extends Exception {
-}
