--- services/ip_manager/src/test/java/com/futurewei/alcor/ipmanager/IpManagerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.futurewei.alcor.ipmanager;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class IpManagerApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
