+++ services/private_ip_manager/src/main/java/com/futurewei/alcor/privateipmanager/entity/Ipv4AddrState.java
@@ -1,0 +1,37 @@
+/*
+Copyright 2019 The Alcor Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+        you may not use this file except in compliance with the License.
+        You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+        Unless required by applicable law or agreed to in writing, software
+        distributed under the License is distributed on an "AS IS" BASIS,
+        WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+        See the License for the specific language governing permissions and
+        limitations under the License.
+*/
+
+package com.futurewei.alcor.privateipmanager.entity;
+
+public enum Ipv4AddrState {
+    ACTIVATED("activated"),
+    DEACTIVATED("deactivated"),
+    FREE("free");
+
+    private String state;
+
+    Ipv4AddrState(String state) {
+        this.state = state;
+    }
+
+    public String getState() {
+        return state;
+    }
+
+    public void setState(String state) {
+        this.state = state;
+    }
+}
