+++ services/private_ip_manager/src/main/java/com/futurewei/alcor/privateipmanager/utils/Ipv4AddrUtil.java
@@ -1,0 +1,50 @@
+/*
+Copyright 2019 The Alcor Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+        you may not use this file except in compliance with the License.
+        You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+        Unless required by applicable law or agreed to in writing, software
+        distributed under the License is distributed on an "AS IS" BASIS,
+        WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+        See the License for the specific language governing permissions and
+        limitations under the License.
+*/
+
+package com.futurewei.alcor.privateipmanager.utils;
+
+public class Ipv4AddrUtil {
+
+    public static boolean formatCheck(String strIp) {
+        if (strIp != null && !strIp.isEmpty()) {
+
+            String regex = "^(1\\d{2}|2[0-4]\\d|25[0-5]|[1-9]\\d|[1-9])\\." +
+                    "(1\\d{2}|2[0-4]\\d|25[0-5]|[1-9]\\d|\\d)\\." +
+                    "(1\\d{2}|2[0-4]\\d|25[0-5]|[1-9]\\d|\\d)\\." +
+                    "(1\\d{2}|2[0-4]\\d|25[0-5]|[1-9]\\d|\\d)$";
+
+            if (strIp.matches(regex)) {
+                return true;
+            } else {
+                return false;
+            }
+        }
+
+        return false;
+    }
+
+    public static long ipToLong(String strIp) {
+        String[] ip = strIp.split("\\.");
+        return (Long.parseLong(ip[0]) << 24) + (Long.parseLong(ip[1]) << 16) + (Long.parseLong(ip[2]) << 8) + Long.parseLong(ip[3]);
+    }
+
+    public static String longToIp(long longIp) {
+        return ((longIp >> 24) & 0xFF) +
+                "." + ((longIp >> 16) & 0xFF) +
+                "." + ((longIp >> 8) & 0xFF) +
+                "." + (longIp & 0xFF);
+    }
+}
