+++ services/private_ip_manager/src/test/java/com/futurewei/alcor/privateipmanager/IpManagerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.futurewei.alcor.privateipmanager;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class IpManagerApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
